# src/main/java/com/demo/employees/domain/Company.java
package com.demo.employees.domain;

import jakarta.persistence.*;
import lombok.*;

import java.util.List;

@Entity
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Company {

    @Setter(AccessLevel.NONE)
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private long id;

    @Column(unique = true)
    private String organisationNumber;
    private String name;

    @OneToMany
    private List<Employee> employees;

}