# src/main/java/com/demo/employees/domain/Employee.java
package com.demo.employees.domain;

import jakarta.persistence.*;
import lombok.*;

@Entity
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Employee {

    @Setter(AccessLevel.NONE)
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private long id;

    @Column(unique= true)
    private String socialSecurityNumber;
    private String firstname;
    private String lastname;

}